+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/dialog/package-info.java
@@ -22,6 +22,7 @@
  * <li>{@link is.codion.swing.common.ui.dialog.ExceptionDialogBuilder#SYSTEM_PROPERTIES}
  * <li>{@link is.codion.swing.common.ui.dialog.ProgressWorkerDialogBuilder#SHOW_DELAY}
  * <li>{@link is.codion.swing.common.ui.dialog.ProgressWorkerDialogBuilder#HIDE_DELAY}
+ * <li>{@link is.codion.swing.common.ui.dialog.LookAndFeelSelectionDialogBuilder#ALLOW_INSTALLED}
  * <li>{@link is.codion.swing.common.ui.laf.LookAndFeelComboBox#INCLUDE_INSTALLED_LOOK_AND_FEELS}
  * </ul>
  */